+++ lib/td_bg_web/swagger_definitions.ex
@@ -213,7 +213,6 @@ defmodule TdBgWeb.SwaggerDefinitions do
             type(:string, "Business Concept type", required: true)
             content(:object, "Business Concept Version content", required: true)
             name(:string, "Business Concept Version name", required: true)
-            description(:object, "Business Concept Version description", required: true)
             last_change_by(:integer, "Business Concept Version last change by", required: true)
             last_change_at(:string, "Business Concept Version last change at", required: true)
             domain(Schema.ref(:DomainRef))
@@ -243,7 +242,6 @@ defmodule TdBgWeb.SwaggerDefinitions do
                 properties do
                   content(:object, "Business Concept Version content")
                   name(:string, "Business Concept Version name")
-                  description(:object, "Business Concept Version description")
                 end
               end
             )
@@ -327,7 +325,6 @@ defmodule TdBgWeb.SwaggerDefinitions do
                   type(:string, "Business Concept type (empty,...)", required: true)
                   content(:object, "Business Concept content", required: true)
                   name(:string, "Business Concept name", required: true)
-                  description(:object, "Business Concept description", required: true)
                   domain_id(:integer, "Business Concept Domain ID", required: true)
                 end
               end