+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-6197] Migrate descriptions to new field in the template
+
 ## [5.20.0] 2023-12-19
 
 ### Changed
 
 - [TD-6215] Use td-core-lib for elastic search and reindex
+
 
 ## [5.19.0] 2023-11-28